reviewing again the EventCoutner stuff... junit are stabilized

+++ drools-framework-runtime-base/src/main/java/org/chtijbug/drools/runtime/RuleBaseSession.java
@@ -73,7 +73,7 @@ public interface RuleBaseSession {
 
     public int getNumberRulesExecuted();
 
-    public int getSessionId();
+    public Long getSessionId();
 
     public Long getRuleBaseID();
